+++ services/presence-service/src/routes/state.routes.js
@@ -12,19 +12,19 @@ async function checkIfStateIsValid(state) {
   }
 }
 
-async function updateUserState(id, state) {
-  try {
-    await redisClient.set(`user:state:${id}`, state);
-    await redisClient.publish(
-      "presence:updates",
-      JSON.stringify({ userId: id, state })
-    );
-  } catch (err) {
-    const error = new Error("Failed to update user state in Redis");
-    error.statusCode = 500;
-    throw error;
-  }
-}
+// async function updateUserState(id, state) {
+//   try {
+//     await redisClient.set(`user:state:${id}`, state);
+//     await redisClient.publish(
+//       "presence:updates",
+//       JSON.stringify({ userId: id, state })
+//     );
+//   } catch (err) {
+//     const error = new Error("Failed to update user state in Redis");
+//     error.statusCode = 500;
+//     throw error;
+//   }
+// }
 
 async function getUserState(id) {
   try {
